Rename/move analysis-related modules to make more sense.
